+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+## Changed
+- Broadcast joining height multiple times #1261
 
 ## [v2.8.2] - 2020-05-26
 ## Added